try to send the pingpong and simplifie the send functions

+++ ESPNOW/TestEspNow_ESP32/main/espnow_example.h
@@ -72,7 +72,7 @@ typedef struct {
     bool broadcast;                       //Send broadcast ESPNOW data.
     bool pingpong;                       //Define the sender and receiver 
     uint8_t state;                        //Indicate that if has received broadcast ESPNOW data or not.
-    uint8_t magic;                       //Magic number which is used to determine which device to send unicast ESPNOW data.
+    uint32_t magic;                       //Magic number which is used to determine which device to send unicast ESPNOW data.
     uint16_t count;                       //Total count of unicast ESPNOW data to be sent.
     uint16_t delay;                       //Delay between sending two ESPNOW data, unit: ms.
     u_int8_t len;                              //Length of ESPNOW data to be sent, unit: byte.
